Create README.md
### Distro Builder
This is a powerful and easy way to build distributions with chroot and Xephyr at its core. A base iso download option is available to work with as well. This script handles everything up to chroot and after leaving the meat of the build to be handled by the pre defined automation script or ones own automation scripts.